+++ src/_template/types.ts
@@ -1,0 +1,6 @@
+// @ts-ignore
+/*
+
+RECORD TYPES HERE
+
+*/