+++ tests/sign.test.ts
@@ -1,20 +1,39 @@
-import test  from 'tape'
+import test from 'tape'
+import { jumpStartNetwork } from '@entropyxyz/sdk/testing'
 
-import { signWithAdapters } from '../src/flows/sign/sign'
-import { setupTest, charlieStashSeed } from './testing-utils'
+import { EntropySign } from '../src/sign/main'
+import { setupTest, eveSeed } from './testing-utils'
 
-test('Sign - signWithAdapter', async (t) => {
-  const { run, entropy } = await setupTest(t, { seed: charlieStashSeed })
+const endpoint = 'ws://127.0.0.1:9944'
+
+test('Sign - signMessageWithAdapters', async (t) => {
+  const { run, entropy } = await setupTest(t, { seed: eveSeed })
+  await run('jump-start network', jumpStartNetwork(entropy))
+  const signService = new EntropySign(entropy, endpoint)
 
   await run('register', entropy.register())
-
-  const signature = await run(
+  const result = await run(
     'sign',
-    signWithAdapters(entropy, { msg: "heyo!" })
+    signService.signMessageWithAdapters({ msg: "heyo!" })
   )
 
-  t.true(signature && signature.length > 32, 'signature has some body!')
-  signature && console.log(signature)
+  t.true(result?.signature?.length > 32, 'signature has some body!')
+  console.log(result)
+
+  t.end()
+})
+
+test('Sign - signMessageWithAdapters (no verifying key)', async (t) => {
+  const { entropy } = await setupTest(t)
+  const signService = new EntropySign(entropy, endpoint)
+
+  const description = 'Unregistered user gets error when they try to sign.'
+  await signService.signMessageWithAdapters({ msg: "heyo!" })
+    .then(() => t.fail(description))
+    .catch((err) => {
+      t.match(err.message, /Cannot read properties of undefined/, description)
+      return
+    })
 
   t.end()
 })